refactor: rename 'ants_at_work_backend' to 'antsback' for more readability
