+++ src/main/java/com/empacoters/antsback/AntsBackApplication.java
@@ -1,13 +1,13 @@
-package com.empacoters.ants_at_work_backend;
+package com.empacoters.antsback;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AntsAtWorkBackendApplication {
+public class AntsBackApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AntsAtWorkBackendApplication.class, args);
+		SpringApplication.run(AntsBackApplication.class, args);
 	}
 
 }
